Separate prepare-activity-log hook and log-activity hook into new files
